+++ elfo/tests/ui/msg_double_wild.rs
@@ -1,0 +1,11 @@
+use elfo::{messages::Terminate, msg, Envelope};
+
+fn test(envelope: Envelope) {
+    msg!(match envelope {
+        Terminate => {}
+        a => {}
+        b => {}
+    });
+}
+
+fn main() {}