+++ elfo/tests/ui/msg_request_syntax_for_regular.rs
@@ -1,0 +1,12 @@
+use elfo::{message, msg, Envelope};
+
+#[message]
+struct SomeEvent;
+
+fn test(envelope: Envelope) {
+    msg!(match envelope {
+        (SomeEvent, token) => {}
+    });
+}
+
+fn main() {}